// ui-next/src/components/features/getStartedSample/types.ts
export enum OperatingSystemEnvironment {
  MAC = "MacOs/Linux",
  WINDOWS = "Windows",
}

export enum CodeLanguage {
  JAVA = "Java",
  PYTHON = "Python",
  GO = "Golang",
  CSHARP = "CSharp",
  JS = "JavaScript",
  CLOJURE = "Clojure",
  GROOVY = "Groovy",
}

export enum JavaLanguageSet {
  GRADLE = "Gradle",
  SPRING_GRADLE = "Spring + Gradle",
}